再提出

+++ src/Review01.java
@@ -20,8 +20,8 @@ public class Review01 {
         int result;
         result = taxMethod(price, taxRate); //taxMethoodを呼び出し
 
-        System.out.println(price + "円の商品の税込価格は" + (price + (int)result) +
-                           "円（消費税は" + (int)result + "円）です。");
+        System.out.println(price + "円の商品の税込価格は" + (price + result) +
+                           "円（消費税は" + result + "円）です。");
         }
 
     public static int taxMethod(int price, int taxRate) {
